test-requirements.txt
testtools
stestr
fixtures
Pillow
